Introduce checked exceptions
For now there is only one native GraphQL exception: `NotFound` but the idea is
to create a GraphQL exceptions hierarchy.

Also accumulation of errors is still not considered, right now
it *short-circuits` when an exception is found. All `Validation` data types I
checked I checked can't be a `Monad`, only `Applicative` so more thought is
required on this.